pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cyborg_rl"
version = "1.0.0"
description = "Production-grade RL system with PMM memory integration"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "CyborgMind Team"}
]
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "gymnasium>=0.29.0",
    "einops>=0.7.0",
    "prometheus-client>=0.19.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]
mamba = [
    "mamba-ssm>=1.2.0",
]
minerl = [
    "minerl>=0.4.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["cyborg_rl*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"
